+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_skeletons.py
@@ -30,7 +30,7 @@ class AllSkeletonsInteractionPlugin(_StageManagerUSDInteractionPlugin):
     ]
     tree: _StageManagerTreePlugin = {"name": "SkeletonGroupsTreePlugin"}
 
-    compatible_trees: list[str] = ["SkeletonGroupsTreePlugin", "PrimGroupsTreePlugin", "VirtualGroupsTreePlugin"]
+    compatible_trees: list[str] = ["SkeletonGroupsTreePlugin", "PrimGroupsTreePlugin"]
     compatible_filters: list[str] = [
         "IgnorePrimsFilterPlugin",
         "IsCaptureFilterPlugin",